move .py files in bot directory
