+++ bot/help_functions.py
